requirements.txt
django
djangorestframework
djangorestframework-jwt
#django-rest-swagger